Data loading function

This function is used to load information on scaffolds into a single object that is used by all subsequent mmgenome functions.

mmload(assembly, coverage, essential)

Arguments

assembly
(required) The assembly.

coverage
(required) A vector with the names of the coverage datasets.

essential
A dataframe with information on essential genes.

pca
Add PCA analyis of tetranucleotide frequencies (default: T)

pca.ncomp
Number of principal components to return (default: 3).

tax
A dataframe with taxonomic classification.

tax.freq
Remove phyla seen less than X times (default: 20).

tax.clean
Clean up some of the phyla names (default: T)

tax.expand
Use class level assignments for specific phyla.

other
A vector with the names of additional datasets.

Value

A list with 2 dataframes: scaffolds and essential (if provided).

Examples

# NOT RUN {
d <- mmload(assembly = assembly,
            coverage = c("A", "B", "C", "D"),
            essential = ess,
            tax = tax,
            tax.expand = "Proteobacteria",
            tax.freq = 50,
            other = "rRNA16S"
            )
# }